Reply Back-end start. Front-end reply show fixed
Back-end route, user/reply, will be made to add replies to clucks on the
server. The clucks replies are stored inside clucks which are stored in
users. To view replies for a cluck, I need to make the user/status/:cluckId
routes. This path will show the cluck with the replies below it, but
this will require another database. The database will give ids for the
path and easy access to those clucks. Without the database, every
execution of the mentioned path will require a for loop to look inside
every users' cluck list and search each until the id is found. This
requires many operations to select the list because the list is in an
object. The memory cost of storing the entire list of users is great.

Another problem is the user objects. They should be stored in seperate
columns if  that improves efficiency. I say if because mongoDB is a
non-relational db unlike sql. The mongoose Schemas may be converted into
a column strcutre/relation in the implementation.After all, the object
schemas approach seems too inefficient to be a provided choice.
I need to research mongoDB.