Reply route added
I added the reply route but did not test it. Cluck.js requires the authUser info passed down to it from Menu. react.createContext() and Context.Providerat the menu level are possible implementations to pass down the token. Once Menu passes the token, the middleware can add the sender info to the cluck.js-type object payload.