Cluck.js reply front-end updated. Providers and context added.
Context was not tested. Reply route not updated. I should create a button to make a cluck at the top of the home page and create a page that displays the cluck and its replies. I can test whether a cluck is added to the reply list by logging the cluck object data in the route. I need to create test_data to simulate a logged in user whose database id is already passed to the route. In production, this is done with the middleware token verifcation.